requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
click==8.1.7
databases==0.9.0
fastapi==0.115.6
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
idna==3.10
protobuf==5.29.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
grpcio-status==1.68.1
googleapis-common-protos==1.66.0
